Project upload
Ransomware has been made that only works on windows for now. Be careful when using the program. Do not run the codes on your own computer without examining them. It is recommended to try it in a virtual machine. The person who wrote the program has no responsibility for the problems that may happen to you.